+++ Backend/Strategies/OpenPlatform/Models/InvestorParameters.cs
@@ -1,9 +1,9 @@
 ﻿namespace Ezbob.Backend.Strategies.OpenPlatform.Models {
     public class InvestorParameters {
         public int InvestorID { get; set; }
-        public double Balance { get; set; }
-        public double DailyAvailableAmount { get; set; }
-        public double WeeklyAvailableAmount { get; set; }
+        public decimal Balance { get; set; }
+        public decimal DailyAvailableAmount { get; set; }
+        public decimal WeeklyAvailableAmount { get; set; }
     }//class InvestorParameters
 
 }//ns
